Add seconds as a counter to give us a unique index
This is because the expanded data table would expand all rows with the
same index and since the time was used as the index and some days have
multiple rides starting at the same time, this made it a problem. This
is a hack, but has worked around the problem.