Issue #18: Expand ride filter toggle buttons to all be same width
By that, I mean that the group should have the same width for each row. Also add
a little bit more padding by putting them in a container so there is some visual
space between the rows.

This change also picks up some autoformatting from the editor that moves a lot
of the CSS to the left column.